refactor(protocol): move SurfaceKey string codec into Adapter namespace
Channels band seam precursor (#499, #503): the pure surface-key string
codec (create/fromChannel/parse + format assertion) is adapter wire
vocabulary, not storage — move it from @openomni/session into the
existing protocol Adapter namespace as Adapter.SurfaceKey. No new
top-level noun is introduced; #499 will carry this codec into the
converged Channel namespace when the band lands.

The session surface-key store keeps only storage semantics
(register/claim/lookup/unregister/listBySession) and imports the codec
for its format invariant via Adapter.SurfaceKey.assertWellFormed — a
single validation authority shared with Adapter.SurfaceKey.create.

All codec callers switch to the protocol import: channel normalizers
and surfaces (killing the dynamic session import in telegram/surface),
ingress bridge, conversation handler, and the kernel session-resolver
(which keeps session imports only for claim/lookup). Pure codec tests
move to packages/protocol/test/adapter-surface-key.test.ts; the session
suite keeps store semantics and gains a claim-format pin.
