+++ apps/server/src/channel/discord/normalizer.ts
@@ -1,5 +1,4 @@
-import type { Adapter } from "@openomni/protocol";
-import { SurfaceKey } from "@openomni/session";
+import { Adapter } from "@openomni/protocol";
 import { normalizeContent } from "../../shared/trigger";
 import type { InboundNormalizer } from "../types";
 import type { DiscordMessage } from "./types";
@@ -26,7 +25,7 @@ export class DiscordNormalizer implements InboundNormalizer<DiscordMessage> {
     content = normalizeContent(content, this.ctx.triggers);
     if (!content) return null;
 
-    const surfaceKey = SurfaceKey.fromChannel({
+    const surfaceKey = Adapter.SurfaceKey.fromChannel({
       surface: "discord",
       namespace: this.ctx.botId,
       kind: isDM ? "dm" : "channel",
