+++ apps/server/src/channel/discord/surface.ts
@@ -1,6 +1,5 @@
-import type { Adapter } from "@openomni/protocol";
-import { Operational, PolicyDecision } from "@openomni/protocol";
-import { Bus, SurfaceKey } from "@openomni/session";
+import { Adapter, Operational, PolicyDecision } from "@openomni/protocol";
+import { Bus } from "@openomni/session";
 import { Dedupe } from "../../shared/dedupe";
 import { DiscordClient } from "./client";
 import { DiscordGateway } from "./gateway";
@@ -78,7 +77,7 @@ export class DiscordAdapter implements Adapter.Surface {
   }
 
   async send(surfaceKey: string, message: Adapter.OutboundMessage): Promise<void> {
-    const parsed = SurfaceKey.parse(surfaceKey);
+    const parsed = Adapter.SurfaceKey.parse(surfaceKey);
     if (!parsed.id) {
       throw new Error(`[discord] surface key missing id: ${surfaceKey}`);
     }
