+++ apps/server/src/channel/github/normalizer.ts
@@ -1,5 +1,4 @@
-import type { Adapter } from "@openomni/protocol";
-import { SurfaceKey } from "@openomni/session";
+import { Adapter } from "@openomni/protocol";
 import { normalizeContent } from "../../shared/trigger";
 import type { GitHubEventContent } from "./types";
 
@@ -16,7 +15,7 @@ export class GitHubNormalizer {
     eventKey: string,
     deliveryId?: string,
   ): Adapter.InboundMessage | null {
-    const surfaceKey = SurfaceKey.fromChannel({
+    const surfaceKey = Adapter.SurfaceKey.fromChannel({
       surface: "github",
       namespace: content.repo,
       kind: "channel",
