+++ packages/protocol/src/adapter/index.ts
@@ -114,4 +114,117 @@ export namespace Adapter {
     registerCommands?(commands: Command[]): Promise<void>;
     onCommand?(handler: CommandHandler): void;
   }
+
+  /**
+   * Pure string codec for surface keys — the wire vocabulary channel
+   * adapters and routing share. Moved here from @openomni/session (#499
+   * precursor): the codec is adapter vocabulary; storage semantics
+   * (register/claim/lookup) stay in the session surface-key store, which
+   * imports this codec for format validation.
+   *
+   * Key format: `<surface-type>:<surface-specific-path>`
+   *
+   * Channel/peer kind encoding:
+   *   - DM:      slack:workspaceA:dm:U123
+   *   - Group:   slack:workspaceA:group:C123
+   *   - Thread:  slack:workspaceA:group:C123:thread:171000
+   *   - Channel: slack:workspaceA:channel:C123
+   *   - TUI:     tui:/Users/ino/Develop/OpenOmni
+   *   - Chat:    telegram:botId:chat:chatId
+   */
+  export namespace SurfaceKey {
+    export type ChannelKind = "dm" | "group" | "channel" | "thread" | "chat";
+
+    export interface ParsedKey {
+      readonly surface: string;
+      readonly namespace: string;
+      readonly kind: ChannelKind | undefined;
+      readonly id: string | undefined;
+      readonly threadId: string | undefined;
+    }
+
+    export interface ChannelDescriptor {
+      /** Surface type (e.g., "slack", "telegram", "tui") */
+      surface: string;
+      /** Namespace/workspace/bot identifier */
+      namespace: string;
+      kind: ChannelKind;
+      id: string;
+      /** Optional thread identifier (creates a sub-key under the channel) */
+      threadId?: string;
+    }
+
+    /**
+     * Assert the surfaceKey wire-format invariant. The single validation
+     * authority: `create` and the session store's register/claim call this.
+     * @returns the key unchanged when well-formed
+     */
+    export function assertWellFormed(key: string): string {
+      if (!key.includes(":")) {
+        throw new Error(
+          `Invalid surfaceKey format: "${key}". Must include surface type prefix (e.g., "slack:...")`,
+        );
+      }
+      return key;
+    }
+
+    /**
+     * Create a surfaceKey from parts.
+     * @param parts - Array of strings to join with colons
+     * @returns Formatted surfaceKey
+     * @throws Error if parts is empty or validation fails
+     */
+    export function create(parts: string[]): string {
+      if (parts.length === 0) {
+        throw new Error("SurfaceKey parts cannot be empty");
+      }
+
+      return assertWellFormed(parts.join(":"));
+    }
+
+    const KNOWN_KINDS: ReadonlySet<string> = new Set<ChannelKind>([
+      "dm",
+      "group",
+      "channel",
+      "thread",
+      "chat",
+    ]);
+
+    export function fromChannel(descriptor: ChannelDescriptor): string {
+      const parts = [descriptor.surface, descriptor.namespace, descriptor.kind, descriptor.id];
+      if (descriptor.threadId) {
+        parts.push("thread", descriptor.threadId);
+      }
+      return create(parts);
+    }
+
+    export function parse(key: string): ParsedKey {
+      const segments = key.split(":");
+      const surface = segments[0] ?? "";
+      const namespace = segments[1] ?? "";
+
+      let kind: ChannelKind | undefined;
+      let id: string | undefined;
+      let threadId: string | undefined;
+
+      for (let i = 2; i < segments.length; i++) {
+        const seg = segments[i];
+        if (seg == null) {
+          continue;
+        }
+        if (KNOWN_KINDS.has(seg)) {
+          if (seg === "thread") {
+            threadId = segments[i + 1];
+            i++;
+          } else {
+            kind = seg as ChannelKind;
+            id = segments[i + 1];
+            i++;
+          }
+        }
+      }
+
+      return { surface, namespace, kind, id, threadId };
+    }
+  }
 }
